split out IterableInAppMessage
